+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog 
 
+## Version 9.0.0 - 2022-04-05
+
+Add support for line items.
+
+- Add optional `fields` to FieldConfig type, allowing nesting FieldConfig for line items
+- BREAKING: `GroundTruth` type has been renamed to `GroundTruthItem`
+- BREAKING: Every occurence of `Array<GroundTruth>` has been made into a new `GroundTruth` type, which is equal to `Array<GroundTruthItem>`
+
 ## Version 8.0.1 - 2022-03-30
 
 - Add trainingId to PredictionResponse type